requirements.txt
pymysql
python-dotenv
fastapi
uvicorn
pydantic
sqlalchemy
sqlglot
openai 
httpx 